Don't derive base classes from QObject
This makes them rather more heavy-weight than necessary. The only reason
they were derived from QObject was to expose them to QML, but so far
this functionality is not actually used. Only Tiled::Map was passed on
in QML and this use-case was replaced by a small wrapper.

In general we now have "Editable" wrappers around these base objects,
which I'd aim to expose in the TiledQuickPlugin somehow, for use by
Tiled Quick.